Pull out MetadataReporter into a separate file.
